centralized skip_build script with real working git commands
